+++ GUI/Actions/HealthCheck.py
@@ -1,0 +1,9 @@
+from Command import Command
+
+class HealthCheck(Command):
+    def __init__(self, receiver):
+        self.receiver = receiver
+        self.message = "Health Check"
+    def execute(self):
+        self.receiver.printmessage(self.message)
+        pass
