+++ GUI/killPythonSessions.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
 
 pkill -x python3
+
+
+Get-Process | Where Name -Like "python" | Stop-Process
