+++ GUI/main.py
@@ -5,19 +5,40 @@ from PyQt5.QtQuick import QQuickView
 from PyQt5.QtQml import qmlRegisterType, QQmlComponent, QQmlEngine
 from PyQt5.QtQuickWidgets import QQuickWidget
 import sys
-from AlbertaLoop_UI import Ui_MainWindow
-import telemetry_module
+from AlbertaLoop_UI2 import Ui_MainWindow
+# import telemetry_module
+from TelemetryModel import TelemetryModel
+from Logic import Logic
+from HealthCheckReq import HealthCheckReq
+from HealthCheckModel import HealthCheckModel
+
+
+from Actions.Launch import Launch
+from Actions.PrepareLaunch import PrepareLaunch
+from Actions.EStop import EStop
+from Actions.Crawl import Crawl
+from Actions.HealthCheck import HealthCheck
+
+
 import time
 from threading import Thread
 from datetime import datetime
 from argparse import ArgumentParser
+from TelemetryReceiver import TelemetryReceiver
+from NetworkComms.udp_module import UDPModule
+from NetworkComms.cmd_transmitter import CmdTransmitter
+class MWindowWrapper(Ui_MainWindow):
 
 
-class Logic(Ui_MainWindow):
+
     def __init__(self, window, ip, port):
         self.setupUi(window)
         self.ip = ip
         self.port = port
+
+        self.command = None
+        self.receivers = []
+
         # -----------------------------------------------------------------
         # Add functionality below!
         # User Added QML Widget for Speed Gauge
@@ -25,91 +46,53 @@ class Logic(Ui_MainWindow):
         self.spedometerWidget.setClearColor(QtCore.Qt.transparent)
         self.spedometerWidget.setResizeMode(QQuickWidget.SizeRootObjectToView)
         self.spedometerWidget.setSource(QUrl("Guage.qml"))
-        self.speed_guage_layout.addWidget(self.spedometerWidget)
+        self.speedGaugeLayout.addWidget(self.spedometerWidget)
+
+        # Connect clicked functions
+        self.launchBtn.clicked.connect(self.launchBtn_clicked)
+        self.healthChkBtn.clicked.connect(self.healthChkBtn_clicked)
+        self.crawlBtn.clicked.connect(self.crawlBtn_clicked)
+        self.prepLaunchBtn.clicked.connect(self.crawlBtn_clicked)
+        self.eStopBtn.clicked.connect(self.eStopBtn_clicked)
+
 
         # logo added
         pixmap = QtGui.QPixmap("img/Albertaloop_logo.png")
-        self.albertaloop_logo.setPixmap(pixmap)
+        self.albertaloopLogo.setPixmap(pixmap)
 
-        # command button connects
-        self.send_command_button.clicked.connect(self.command_button_clicked)
-        self.send_command_button.clicked.connect(self.command_button_input)
 
-        # Emergency button and Simulation button connects
-        self.estop_button.clicked.connect(self.e_stop_button_clicked)
-        self.simulation_button.clicked.connect(self.simulation_button_clicked)
+    # Clicked function definitions
+    def launchBtn_clicked(self):
+        print("Launch button pressed")
+        self.setCommand(Launch(self.receivers[0]))
+        self.executeCommand()
+    def healthChkBtn_clicked(self):
+        print("Health check button pressed")
+        self.setCommand(HealthCheck(self.receivers[1]))
+        self.executeCommand()
+    def crawlBtn_clicked(self):
+        self.setCommand(Crawl(self.receivers[0]))
+        self.executeCommand()
+        print("Crawl button pressed")
+    def prepLaunchBtn_clicked(self):
+        self.setCommand(PrepareLaunch(self.receivers[0]))
+        self.executeCommand()
+        print("Prepare Launch button pressed")
+    def eStopBtn_clicked(self):
+        self.setCommand(EStop(self.receivers[0]))
+        self.executeCommand()
+        print("Estop button pressed")
 
-        thread = Thread(target=self.getTelemetryData)
-        thread.start()
-
-    def getTelemetryData(self):
-        telemetry_manager = telemetry_module.TelemetryManager(self.ip, self.port)
-        start_time = time.time()
-        self.current_state_ind_2.setText("Active Connection")
-        self.pod_connect_ind.setText("Active Connection")
-        self.thing_2_label.setText("Stripe Count")
-
-        while True:
-            print("Team ID: ", telemetry_manager.get_team_id())
-            print("Status: ", telemetry_manager.get_status())
-            print("Acceleration: ", telemetry_manager.get_acceleration())
-            print("Velocity: ", telemetry_manager.get_velocity())
-            print("Position: ", telemetry_manager.get_position())
-            print("Battery Voltage: ", telemetry_manager.get_battery_voltage())
-            print("Battery Current: ", telemetry_manager.get_battery_current())
-            print("Battery Temperature: ", telemetry_manager.get_battery_temperature())
-            print("Pod Temperature: ", telemetry_manager.get_pod_temperature())
-            print("Stripe Count: ", telemetry_manager.get_stripe_count())
-            print("Highest Velocity: ", telemetry_manager.get_highest_velocity())
-            print("\n")
-
-            self.packetTextBrowser.append(
-                "Team ID: " + str(telemetry_manager.get_team_id())
-            )
-            self.packetTextBrowser.append(
-                "Status: " + str(telemetry_manager.get_status())
-            )
-            self.packetTextBrowser.append(
-                "Acceleration: " + str(telemetry_manager.get_acceleration())
-            )
-            self.packetTextBrowser.append(
-                "Velocity: " + str(telemetry_manager.get_velocity())
-            )
-            self.packetTextBrowser.append(
-                "Position: " + str(telemetry_manager.get_position())
-            )
-            self.packetTextBrowser.append(
-                "Battery Voltage: " + str(telemetry_manager.get_battery_voltage())
-            )
-            self.packetTextBrowser.append(
-                "Battery Current: " + str(telemetry_manager.get_battery_current())
-            )
-            self.packetTextBrowser.append(
-                "Battery Temperature: "
-                + str(telemetry_manager.get_battery_temperature())
-            )
-            self.packetTextBrowser.append(
-                "Pod Temperature: " + str(telemetry_manager.get_pod_temperature())
-            )
-            self.packetTextBrowser.append(
-                "Stripe Count: " + str(telemetry_manager.get_stripe_count())
-            )
-            self.packetTextBrowser.append(
-                "Highest Velocity: " + str(telemetry_manager.get_highest_velocity())
-            )
-            self.packetTextBrowser.append("\n")
-            self.packetTextBrowser.moveCursor(QtGui.QTextCursor.End)
-
-            self.battery_1_temp.setText(str(telemetry_manager.get_battery_voltage()))
-            self.position_ind.setText(str(telemetry_manager.get_position()))
-            self.bat_1_volt.setText(str(telemetry_manager.get_battery_voltage()))
-            # self.progressBar.setProperty("value", telemetry_manager.get_position())
-            elapsed_time = time.time() - start_time
-            self.time_elapsed_label.setText("Time Elapsed: %.2f" % elapsed_time)
-            # self.self.spedometerWidget.engine().setContextProperty('gauge_value', telemetry_manager.get_velocity())
-            time.sleep(1)
-
-    # functionality definitions
+    # Command Pattern definitions
+    def setCommand(self, command):
+        self.command = command
+    def executeCommand(self):
+        self.command.execute()
+    def setReceivers(self, rcv1, rcv2):
+        # Receiver 1 is Logic
+        self.receivers.append(rcv1)
+        # Receiver 2 is HealthChkReq
+        self.receivers.append(rcv2)
 
     def command_button_clicked(self):
         # TODO send general command to pod
@@ -130,7 +113,72 @@ class Logic(Ui_MainWindow):
     def simulation_button_clicked(self):
         # TODO luanch new window to start simulation testing on pod
         print("Entering simulation")
+    #updates label colors if state is not equal to current_state
+    def updateCurrentState(state):
+        current_state = "fault"
+        if state== current_state:
+            True
+        elif state!= current_state:
+            current_state=state
+            if state== 'fault':
+                self.label12.setStylesheet("background-color: red")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'safe':
+                self.label12.setStylesheet("background-color: gray")
+                self.label11.setStylesheet("background-color: #89CFF0")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'ready':
+                self.label12.setStylesheet("background-color: gray")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: green")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'launch':
+                self.label12.setStylesheet("background-color: #89CFF0")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: green")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'coast':
+                self.label12.setStylesheet("background-color: gray")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: green")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'break':
+                self.label12.setStylesheet("background-color: gray")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: yellow")
+                self.label6.setStylesheet("background-color: gray")
+            if state== 'crawl':
+                self.label12.setStylesheet("background-color: gray")
+                self.label11.setStylesheet("background-color: gray")
+                self.label10.setStylesheet("background-color: gray")
+                self.label9.setStylesheet("background-color: gray")
+                self.label8.setStylesheet("background-color: gray")
+                self.label7.setStylesheet("background-color: gray")
+                self.label6.setStylesheet("background-color: yellow")
 
+        def notify(state):
+            updateCurrentstate(state)
 
 if __name__ == "__main__":
     parser = ArgumentParser(description="Albertaloop GUI launch")
@@ -144,7 +192,33 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     app = QApplication(sys.argv)
+
+    # Model Classes
+    TelemetryModel = TelemetryModel()
+    HealthCheckModel = HealthCheckModel()
+    TelemetryReceiver = TelemetryReceiver()
+    CmdTransmitter = CmdTransmitter()
+
+
+    # Controller Classes
+    Logic = Logic(TelemetryModel,CmdTransmitter)
+    TelemetryReceiver.setDataModel(TelemetryModel)
+    HealthCheckReq = HealthCheckReq(HealthCheckModel)
+
+    # TelemetryReceiver.start()
+    # View Classes
     MainWindow = QMainWindow()
-    ui = Logic(MainWindow, args.server_ip, args.server_port)
+    mWindowWrapper = MWindowWrapper(MainWindow, args.server_ip, args.server_port)
+    mWindowWrapper.setReceivers(Logic, HealthCheckReq)
+    TelemetryModel.attach(mWindowWrapper)
+    
+    UDPModule = UDPModule("127.0.0.1", 4000, 3000,
+                          TelemetryReceiver, CmdTransmitter)
+    # TelemetryModel.recieveData([9]*10)
+    # TelemetryModel.notify()
+    # TelemetryModel.update([0.123]*10)
+    
+    
+    
     MainWindow.show()
     sys.exit(app.exec_())
